+++ tests/models.py
@@ -1,7 +1,7 @@
 """Test models."""
 
 from datetime import date
-from typing import Annotated
+from typing import Annotated, Optional
 
 from pydantic import BaseModel, PlainSerializer
 
@@ -21,13 +21,14 @@ class DeezerSong(BaseModel):
     DURATION: Annotated[int, PlainSerializer(str)]
     ART_NAME: str
     SNG_TITLE: str
-    VERSION: str | None = None
+    VERSION: Optional[str] = None
     ALB_TITLE: str
     ALB_PICTURE: str
     PHYSICAL_RELEASE_DATE: Annotated[date, PlainSerializer(str)]
     FILESIZE_MP3_128: Annotated[int, PlainSerializer(str)]
     FILESIZE_MP3_320: Annotated[int, PlainSerializer(str)]
     FILESIZE_FLAC: Annotated[int, PlainSerializer(str)]
+    FALLBACK: "Optional[DeezerSong]" = None
 
 
 class DeezerSongResponse(BaseDeezerGWResponse):