+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Add fallback track support
 - CLI: add track release date to outputs
 
 ## [0.5.2] - 2025-11-04